requirements.txt
django
djangorestframework
bcrypt
PyJWT
pymongo